webapp/requirements.txt
streamlit
python-dotenv
psycopg2-binary
sqlalchemy
pandas
yfinance
sqlglot

langchain-community
langchain-core
langchain-mistralai
langgraph